parkalot-func/requirements.txt

azure-functions
playwright==1.41.0
twilio==8.10.0


